src/applications/requirements.txt
attrs==22.2.0
backoff==2.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Deprecated==1.2.14
docopt==0.6.2
exceptiongroup==1.1.0
Flask==2.2.3
Flask-Cors==4.0.0
Flask-MySQLdb==1.0.1
googleapis-common-protos==1.61.0
grpcio==1.59.2
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector==2.2.9
mysqlclient==2.1.1
opentelemetry-api==1.21.0
opentelemetry-distro==0.41b0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-instrumentation==0.41b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
packaging==23.0
pipreqs==0.4.11
pluggy==1.0.0
protobuf==4.25.0
PyJWT==2.6.0
pytest==7.2.1
pytest-mock==3.12.0
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
thrift==0.16.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==1.26.14
Werkzeug==2.2.3
wrapt==1.15.0
yarg==0.1.9
zipp==3.17.0
